Update vars.tf
Updated from 2.8.1 to 2.10.0, but the canned responses seemed to have disappeared after the update. Reverting back to 2.8.1 to see if this fixes the problem until a permanent solution with 2.10.0 can be found.

NOTE:
Perhaps the issue was that I didn't SSH into the new AWS container that the new Chatwoot instance was instantiated on and run the docker exec and bundle exec commands. I didn't do this because I don't have things setup on my end for me to SSH in.